+++ chatbot-core/backend/app/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "1.0.0"
