+++ chatbot-core/backend/app/models/api.py
@@ -1,10 +1,11 @@
 from pydantic import BaseModel
 from typing import Optional, Any
 
-from utils.error_handler import ErrorCodesMappingNumber
+from app.utils.error_handler import ErrorCodesMappingNumber
 
 
 class APIResponse(BaseModel):
+    status_code: int = 200
     message: str = "Success"
     headers: Optional[Any] = None
     data: Optional[Any] = None
