+++ chatbot-core/backend/app/settings/__init__.py
@@ -1,0 +1,8 @@
+from .constants import Constants
+from .secrets import Secrets
+
+
+__all__ = [
+    "Constants",
+    "Secrets",
+]
