+++ chatbot-core/backend/app/utils/api_response.py
@@ -1,6 +1,7 @@
 from typing import Optional, Any
+from fastapi import status
 
-from models.api import APIResponse
+from app.models.api import APIResponse
 
 
 class BackendAPIResponse:
@@ -12,6 +13,10 @@ class BackendAPIResponse:
         )
 
         self.resp = resp
+
+    def set_status_code(self, status_code: int) -> "BackendAPIResponse":
+        self.resp.status_code = status_code
+        return self
 
     def set_message(self, message: str) -> "BackendAPIResponse":
         """
